+++ packages/plugin-rsc/e2e/basic.test.ts
@@ -1,13 +1,13 @@
 import { createHash } from 'node:crypto'
 import { readFileSync } from 'node:fs'
 import { type Page, expect, test } from '@playwright/test'
-import { type Fixture, setupIsolatedFixture, useFixture } from './fixture'
-import { expectNoReload, testNoJs, waitForHydration } from './helper'
-import path from 'node:path'
-import os from 'node:os'
-
-// TODO: parallel?
-// TODO: all tests don't need to be tested in all variants?
+import { type Fixture, useFixture } from './fixture'
+import {
+  expectNoPageError,
+  expectNoReload,
+  testNoJs,
+  waitForHydration,
+} from './helper'
 
 test.describe('dev-default', () => {
   const f = useFixture({ root: 'examples/basic', mode: 'dev' })
@@ -46,33 +46,9 @@ test.describe('build-default', () => {
   defineTest(f)
 })
 
-test.describe(() => {
-  // disabled by default
-  if (process.env.TEST_ISOLATED !== 'true') return
-
-  // use RUNNER_TEMP on Github Actions
-  // https://github.com/actions/toolkit/issues/518
-  const tmpRoot = path.join(
-    process.env['RUNNER_TEMP'] || os.tmpdir(),
-    'test-vite-rsc',
-  )
-  test.beforeAll(async () => {
-    await setupIsolatedFixture({ src: 'examples/basic', dest: tmpRoot })
-  })
-
-  test.describe('dev-isolated', () => {
-    const f = useFixture({ root: tmpRoot, mode: 'dev' })
-    defineTest(f)
-  })
-
-  test.describe('build-isolated', () => {
-    const f = useFixture({ root: tmpRoot, mode: 'build' })
-    defineTest(f)
-  })
-})
-
 function defineTest(f: Fixture) {
   test('basic', async ({ page }) => {
+    using _ = expectNoPageError(page)
     await page.goto(f.url())
     await waitForHydration(page)
   })
@@ -818,6 +794,14 @@ function defineTest(f: Fixture) {
     )
   })
 
+  test('transitive cjs dep', async ({ page }) => {
+    await page.goto(f.url())
+    await waitForHydration(page)
+    await expect(page.getByTestId('transitive-cjs-client')).toHaveText(
+      'ok:browser',
+    )
+  })
+
   test('use cache function', async ({ page }) => {
     await page.goto(f.url())
     await waitForHydration(page)
@@ -939,9 +923,7 @@ function defineTest(f: Fixture) {
     expect(errors).toMatchObject([
       {
         message: expect.stringContaining(
-          f.mode === 'dev'
-            ? `Hydration failed because the server rendered HTML didn't match the client.`
-            : `Minified React error #418`,
+          f.mode === 'dev' ? `Hydration failed` : `Minified React error #418`,
         ),
       },
     ])